refactor: seperate view and controller in Purchase
+ ProductPurchase.js
Controller

+ ProductPurchaseTemplate.js
View
